backend/requirements.txt
Flask==3.0.0
selenium==4.19.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
splinter==0.19.0
beautifulsoup4==4.12.3
pandas==2.2.1
matplotlib==3.8.3
chromedriver-autoinstaller==0.6.0
Flask-CORS==4.0.0
